requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
duckdb==1.3.0
pyinstaller
frida==16.7.19
frida-tools
structlog>=24.1.0
python-json-logger>=2.0.0
python-logging-loki>=0.3.1
influxdb-client
PyYAML>=6.0
requests>=2.25.0
rich>=13.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
psutil>=5.9.0
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0